Move test classes to a Tests namespace
